requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
timeboard==0.2.3
Werkzeug==1.0.1
